nspawn: fix loading symlinked directoris via --machine
The function chase_symlinks_and_update() is used to find the path symlinks in
/var/lib/machines/ point at. The end of the path of a directory used to have
a trailing slash. The lock file creation code would then run the GNU version
of basename() on the path. This resulted in an invalid argument error.

Strip the trailing slash of directories in chase_symlinks_and_update().

Fixes #8544