meson: sysctl.d/ and other early boot configs should be installed to /lib
and not /usr/lib.  E.g. systemd-sysctl.service has DefaultDependencies=no,
and no explicit dependency on /usr.  Closes #8481.

* binfmt.d
* modules-load.d
* sysctl.d
* sysusers.d
* systemd/system-environment-generators
* tmpfiles.d

Scope
-----

This would affect setups which start systemd without /usr pre-mounted (in
initramfs).  While we claimed systemd itself supported such setup, nowadays
this should be pretty rare.  Debian stretch has officially withdrawn
support for it.

The problem is more that Debian is currently still carrying the /usr split.
This makes systemd's failure to support mounting /usr quite visible...
when you look at the installed paths, or the doctored manpages from the
Debian systemd package.

Given that we make this distinction *at all* between installing things to
`prefixdir` and `rootprefixdir`, let's make a story for readers which is
consistent, one which doesn't look like there's a bug, without assuming
they are up to date on the latest developments.

How we avoid breaking existing setups
-------------------------------------

No extra migration work should be necessary.  While upstream manpages
pretend the /usr split doesn't exist, in fact when we are built with
split_usr=true we are looking in both /lib AND /usr/lib for config files.
So when /usr is pre-mounted, we will read it opportunistically.
(Also, we do so in such a way that allows a graceful migration.  It
shouldn't matter if /lib/foo.conf and /usr/lib/foo.conf end up being two
different aliases of the same file, because only one of them will be read).

A change may be desirable in the Debian package, to include these new
directories in the `sed` command which doctors their man pages :).

Tested
------

On a small nspawn, without /lib/ being a symlink to /usr/lib, boot remains
successful with no error messages in `journalctl -b`.