journal-file: detect and propagate fsync() errors
An fsync_error member has been added to JournalFile for caching errno
when an fsync on the underlying journal fd fails.
A helper function is used to maintain this cached errno state when
calling fsync() outside of the asynchronous offlining state machine.

The offline thread now has an OFFLINE_FAILED state synonymous with
OFFLINE_DONE it may enter upon seeing an error from fsync().  If the
first fsync() call of the pair of required fsyncs fails, the second
one accompanying the header switch to archived/offline no longer
occurs so the journal may be detected as online/corrupt when affected
by errors.

A status variable is used in the offlining state machine to store a
failed fsync()'s errno value.  Any subsequent attempt to perform an
fsync in the offlining state machine is prevented when status < 0, and
treated as if the fsync had failed with its errno already in status,
(re)attempting transitioning to the terminal OFFLINE_FAILED state.

All return paths from the offlining state machine return the status
value.

Synchronous offline operations which don't involve the threaded
offline machinery now receive the status return value directly, and
propagate it up the stack as any other errors.

The asynchronous offline thread now returns the status returned from
the offlining state machine, its sole function.  The offline thread
joining has been changed to retrieve the thread's status, and its
value treated as an error value to return when < 0.  Care has been
taken to not short-circuit processing of bus errors in the mmap cache
when the offlining thread returns an error.

See [1] for more information.

[1] https://lists.freedesktop.org/archives/systemd-devel/2018-April/040610.html