udev: lock symlink directories in link_update()
When uevents are processed in parallel, in particular during coldplug,
it can happen that several events for the same symlink arrive at the
same time (e.g. SCSI disk and for a multipath DM device for that
disk). This opens up a race window where a higher-priority link target
may not be visible under /run/udev/links/ yet when a lower-priority
one is processed, leading to the creation of the wrong symlink.

The recently introduced "multiple symlinks" test from the udev test suite
reveals this race condition, most easily if run on a system with many cores.

We have to avoid this by locking the directory under /run/udev/links/.
I have extensively studied alternative approaches; nothing besides
locking avoids the said race condition reliably.

Initial attempts to do this with systemd's make_lock_file_for() API
revealed another problem: In certain situations, there may be a lot of
contenders for a given symlink (for example, on various distributions, almost
all GPT partitions have the partition labels "primary" or "logical" due
to a bug in past parted releases - I have seen systems with literally
thousands of contenders for the /dev/disk/by-partlabel/primary symlink).
As of 4.18, the kernel implementation of fcntl(F_OFD_SETLKW) is inefficient,
leading to busy-waiting in spin-locks and thus very bad scaling when
many contending processes compete for a file lock. The kernel implementation
can be improved and patches are under discussion, but it will take a long
time until that kernel-side improvement is generally available.

To avoid this problem, this patch doesn't use make_lock_file_for().
Rather, it uses a set of SysV semaphores, where the index into the semaphore
set is derived from the symlink name using udev's built-in hash function.
This yields good collision avoidance for different symlink names and performs
way better than fcntl(F_OFD_SETLKW) with many contenders.