udev: net_id - handle Hyper-V synthetic NICs
Hyper-V has an abstract bus, which gets renumbered on guest startup. So
let's use the device GUIDs, which can be retrieved from the 'device_id'
sysfs attribute.

In order to fit the network interface name length limit of 15
characters, let's use only the last part of the GUID, which has length
12, added to the "enh" prefix to get to the maximum of 15.