RFC: Add a new command to let systemd resume from hibernate
The idea would be that systemd-sleep would be placed into
the initramfs and called after swap got mounted.

It would look up which swap location was preferred and check
for a header indicating it should try to resume.

Unfortunately this is predicated upon the kernel exporting
the format. An alternative can be to just "try" on every
boot to resume (the kernel will then check for this header
silently ignore the failure unless configured to output debug level)